fix: add new exclude-rule to golangci-lint config; move lint task to be after swag task
An issue was occuring wherein the swag formatter ('swag fmt' command) was clashing with the formatting that the goimports linter desired. Since linting was happening before the swag formatting, this was not showing up during some pcc task runs. Task order has been rearranged to lessen chance of similar occurences in the future. An exclude-rule was set on the config to ensure the same issue no longer occurs. The downside is that the rule will no longer be enforced for the project, however this rule has not been triggered before and is deemed to be low risk.