fix: remove dupl linter
This linter was raising an error due to the duplicate user and product handler functions, despite the utilisation of generic handlers. In order to remove this error you could call the generic handlers from the router method directly, however this would reduce code readability and remove the ability to have swagger annotations for each handler.